Move useProductTypeSelector into the shared folder
