+++ plugins/woocommerce/client/blocks/assets/js/blocks/add-to-cart-with-options/components/toolbar-type-product-selector-group/index.tsx
@@ -16,7 +16,7 @@ import {
 /**
  * Internal dependencies
  */
-import useProductTypeSelector from '../../hooks/use-product-type-selector';
+import { useProductTypeSelector } from '../../../../shared/stores/product-type-template-state';
 
 export default function ToolbarProductTypeGroup() {
 	const {
