+++ plugins/woocommerce/client/blocks/assets/js/blocks/add-to-cart-with-options/plugins/index.tsx
@@ -13,7 +13,7 @@ import {
 /**
  * Internal dependencies
  */
-import useProductTypeSelector from '../hooks/use-product-type-selector';
+import { useProductTypeSelector } from '../../../shared/stores/product-type-template-state';
 
 function ProductTypeSwitcher() {
 	const { productTypes, current, set } = useProductTypeSelector();
