+++ plugins/woocommerce/client/blocks/assets/js/blocks/add-to-cart-with-options/types.ts
@@ -1,8 +1,3 @@
-export type ProductTypeProps = {
-	slug: string;
-	label: string;
-};
-
 export type QuantitySelectorStyleProps = 'input' | 'stepper';
 
 export interface Attributes {
