+++ plugins/woocommerce/client/blocks/assets/js/blocks/add-to-cart-with-options/variation-selector/edit.tsx
@@ -8,7 +8,7 @@ import { useProductDataContext } from '@woocommerce/shared-context';
 /**
  * Internal dependencies
  */
-import useProductTypeSelector from '../hooks/use-product-type-selector';
+import { useProductTypeSelector } from '../../../shared/stores/product-type-template-state';
 import { ATTRIBUTE_ITEM_TEMPLATE } from './attribute-item-template/constants';
 
 interface Attributes {
