+++ plugins/woocommerce/client/blocks/assets/js/shared/stores/product-type-template-state/types.ts
@@ -1,0 +1,4 @@
+export type ProductTypeProps = {
+	slug: string;
+	label: string;
+};
